Move server metric reporting to separate class
